[김예림] 로그인 개선

+++ src/components/login/Logout.jsx
@@ -1,35 +1,57 @@
 import React from 'react';
-import { NavLink } from 'react-router-dom';
+import { NavLink, useNavigate } from 'react-router-dom';
 import axios from 'axios';
+import styled from 'styled-components';
 
 const Logout = () => {
 
-    const auth = window.localStorage.getItem("Authorization")
-    const token = localStorage.getItem("Refresh-Token")
-
+    const Authorization = localStorage.getItem('Authorization');
+    const navigate = useNavigate();
     const onLogin= async() => {
-
-        const response = await axios.get("/user/logout",{
-            headers:{
-              "Refresh-Token": token
+        const contest = window.confirm("정말 로그아웃 하실건가요?");
+        if(contest === true){
+            const Refreshtoken = localStorage.getItem('RefreshToken');
+            const headers = {
+                'Content-Type': 'application/json',
+                RefreshToken: Refreshtoken
             }
-          })
+            await axios.delete("http://15.164.170.89/user/logout",
+                {headers:headers},{})
 
-        window.localStorage.clear();
+            window.localStorage.clear();
+            alert('로그아웃 되셨습니다.')
+            navigate("/")
+        }
+        else if(contest === false){
+            return
+        }
     }
     return (
         <div>
-            { auth ===null ?
-            <div><NavLink to = "/login">로그인하기</NavLink></div>
+            { Authorization ===null ?
+            <LoginDiv><NavLink to = "/login">로그인하기</NavLink></LoginDiv>
             :
-            <div
+            <LoginDiv
             className='logout_button'
             onClick={onLogin}>
             로그아웃 하기
-            </div>
+            </LoginDiv>
         }
         </div>
     );
 };
 
 export default Logout;
+
+const LoginDiv = styled.div `
+    text-decoration: none;
+    a{
+        text-decoration: none;
+        color: #000c2d;
+        transition: all .3s;
+        &:hover{
+            color:  ${(props) => props.theme.lightmode.color_brand};
+        }
+    }
+    cursor: pointer;
+`
